On average I think we're losing maybe 15-30ms by not using eval, however I think there's more room to optimise here (seeing some deopts in both cases, e.g. not a Smi and Unexpected Object). Also util.format is unoptimizable so we might want to rewrite that part too. 

The trade off here is using eval vs limited args and possibly up to 30ms hit (quite a lot - could be optimized more though - also we're still miles ahead). Not using eval may help adoption. 

```sh
with eval:
benchPino*10000: 467.518ms
benchPinoObj*10000: 515.426ms
benchPino*10000: 408.458ms
benchPinoObj*10000: 500.840ms

without:
benchPino*10000: 461.231ms
benchPinoObj*10000: 525.623ms
benchPino*10000: 440.931ms
benchPinoObj*10000: 513.321ms
```