@mcollina 2016-02-22
I find the eval thing much more readable.

Also, are unit tests passing? I think `util.format` prints `undefined` for every passed arg that is `undefined`, at least on node v4.