@davidmarkclements 2016-02-22
yes - all tests pass - when we set params.length to arguments.length - base it cuts down the params array to actual inputs length and discards the rest

certainly there's things we can do to make it more readable - also is it possible that eval'ing strings is only more readable to you because you wrote it? :{P