@mcollina 2016-02-22
Agreed, I've refactored this a bit, and added you as a contributor :D.