@davidmarkclements 2016-02-22
sweet - so one other thing - now you've refactored and using closure instead of func-object the function no longer needs to be called "plog" - if you make it anonymous then at least the stack traces can infer function name from object key instead of all logger types being called plog