@davidmarkclements 2016-02-22
https://github.com/mcollina/pino/commit/ea686df747154c4dc9a6a84c6aa492a3d88a0363